+++ CHANGES.md
@@ -4,7 +4,7 @@ bandwidth-go changelog
 [2023-12-21] Version 0.0.2
 ---------------------------
 - Remove unmatched verbs and nouns.
-
+- Add PlayAudio verb.
 
 [2023-12-21] Version 0.0.1
 ---------------------------